some refactors, remove c++11 use

+++ src/factorial.h
@@ -5,7 +5,7 @@
 #ifndef __RAMANUJANFACTORIAL__
 #define __RAMANUJANFACTORIAL__
 
-inline float _ramanujan(float n) {
+inline float _ramanujan(int n) {
   float f = sqrt(M_PI) * pow(n, n) * exp(-n) * pow((8.0 * pow(n, 3.0) + 4.0 * pow(n, 2.0) + n + 1.0 / 30.0), 1.0 / 6.0);
   return f;
 }
@@ -16,7 +16,7 @@ inline float _ramanujan(float n) {
 #ifndef __STIRLINGFACTORIAL__
 #define __STIRLINGFACTORIAL__
 
-inline float _stirling(float n) {
+inline float _stirling(int n) {
   float f = sqrt((2.0 * n + 1.0 / 3.0) * M_PI) * pow(n, n) * exp(-n);
   return f;
 }
@@ -27,7 +27,7 @@ inline float _stirling(float n) {
 #ifndef __STIRLINGLNFACTORIAL__
 #define __STIRLINGLNFACTORIAL__
 
-inline float _stirlingln(float n) {
+inline float _stirlingln(int n) {
   float f = n * log(n) - n + (1.0 / 6.0) * log(n * (1.0 + 4.0 * n * (1.0 + 2.0 * n))) + 0.5 * log(M_PI);
 
   return f;
@@ -39,10 +39,10 @@ inline float _stirlingln(float n) {
 #ifndef __FACTORIAL__
 #define __FACTORIAL__
 
-inline long int _factorial(int n) {
+inline long int _factorial(long int n) {
   long int fac = 1;
 
-  for (int i = n; i >= 1; i--) {
+  for (long int i = n; i >= 1; i--) {
     fac = fac * i;
   }
 
@@ -50,4 +50,3 @@ inline long int _factorial(int n) {
 }
 
 #endif
-
